+++ cell_installer.go
@@ -11,7 +11,7 @@ import (
 	"bytes"
 	"strings"
 	"github.com/pkg/errors"
-	"nano/framework"
+	"github.com/project-nano/framework"
 	"github.com/vishvananda/netlink"
 	"bufio"
 	"net"